fix signed-ness in comparison
git-svn-id: https://nocs.stanford.edu/cgi-bin/svn.cgi/booksim2.0/branches/classes@4618 ce7aba0d-9f81-40a6-a495-cecae84b0515

+++ src/workload.cpp
@@ -479,7 +479,7 @@ NetraceWorkload::~NetraceWorkload()
 
 void NetraceWorkload::_refill()
 {
-  while((_limit < 0) || (_count < _limit)) {
+  while((_limit < 0) || (_count < (unsigned long long int)_limit)) {
     _next_packet = nt_read_packet(_ctx);
     if(!_next_packet) {
 #ifdef DEBUG_NETRACE
